moved all espresso options to espresso_options.py
