--- appium/options/android/espresso/show_gradle_log_option.py
@@ -1,48 +1,0 @@
-# Licensed to the Software Freedom Conservancy (SFC) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The SFC licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-from typing import Optional
-
-from appium.options.base_options_descriptor import OptionsDescriptor
-from appium.options.common.supports_capabilities import SupportsCapabilities
-
-
-class ShowGradleLogOption(SupportsCapabilities):
-    SHOW_GRADLE_LOG = "showGradleLog"
-    show_gradle_log = OptionsDescriptor[Optional[bool], bool](SHOW_GRADLE_LOG)
-    """
-    Whether to include Gradle log to the regular server logs while
-    building Espresso server. false by default.
-
-    Usage
-    -----
-    - Get
-        - `self.show_gradle_log`
-    - Set
-        - `self.show_gradle_log` = `value`
-    
-    Parameters
-    ----------
-    `value`: `bool`
-
-    Returns
-    -------
-    - Get
-        - `Optional[bool]`
-    - Set
-        - `None`
-    """
